Add DON price proxy

+++ contracts/OGS/intf/IOGSPriceProxy.sol
@@ -3,7 +3,7 @@
 pragma solidity 0.6.9;
 
 interface IOGSPriceProxy {
-    function latestPrice() external view returns (uint128);
+    function latestPrice() external view returns (uint256);
 
     function decimals() external view returns (uint8);
 }
